refactor(request): move the device-inventory context out of core
`device-inventory-context.ts` is an AsyncLocalStorage holder: it imports only
`node:async_hooks`, kernel errors, and contracts types, and nothing from `core`.
`daemon-modularity.ts` already describes `src/request/` as "request-global daemon
plumbing (progress sinks, cancellation, AsyncLocalStorage)", and `request/progress.ts`
is the same shape — a `withX`/accessor pair over one store.

Living in `core` (rank 2) made the platform-runtime composition root reach up to
rank 2 for `listLocalDeviceInventory`. Injecting the lookup would not have fixed
that: the only caller, `platform-runtime-operation-host.ts`, is in the same zone,
so the edge would move rather than disappear. Relocating the module does remove it,
and every other consumer — daemon (rank 4) and `core/dispatch-resolve.ts` (rank 2) —
now reaches down instead of sideways or up.

Pure rename; the module body is unchanged and 14 import specifiers are retargeted.
R17's `DEVICES_INVENTORY_IMPORT_SOURCES` pin follows the module: the rule still
asserts the devices handler imports the neutral gateway, does not shadow it, and
calls it.

This clears the last composition-root upward value edge that was not deliberate.
Confirmed with `pnpm depgraph`: those edges drop from 3 to 2, and the two that
remain are the documented ones — `provider-limrun-runtime.ts` (an in-file comment
marks it a deliberate static seam) and `runtime.ts -> commands/index.ts`
(`bindCommands` is the aggregate binder). Cycle, back-edge and R6 counts unchanged.
