+++ src/__tests__/test-utils/device-inventory-gateways.ts
@@ -12,7 +12,7 @@ import type {
 } from '@agent-device/contracts/platform';
 import { createPlatformModuleRegistry } from '@agent-device/contracts/platform-module';
 import type { DeviceInfo, Platform } from '@agent-device/kernel/device';
-import { withDeviceInventoryContext } from '../../core/device-inventory-context.ts';
+import { withDeviceInventoryContext } from '../../request/device-inventory-context.ts';
 import { createComposedDeviceInventoryGateways } from '../../platform-runtime-device-inventory.ts';
 
 type TestDeviceInventoryOptions = Readonly<{
